Improved building hierarchy (read desc)
Improved building hierarchy substantially by creating a buildable gateway called Buildables.cs

This is still a WIP and the setup is messy, but essentially instead of scripts directly accessing the building SO for data, they now run through a buildable master class that contains all the runtime data that building would need.

This is much better then before since SO's we're essentially non-tangible, meaning there was no good way to update that info based on research values, price values, and so on.

Again, this is still a WIP, but it is a step in the right direction that will make applying runtime data much easier!